Save and load methods is now in Service class
